+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.0.3
+* Fixed [#1](https://github.com/vanelizarov/flutter_macos_webview/issues/1): `parentViewController` now returns Flutter view's controller instead of `
+NSApp.keyWindow.contentViewController`
+
 ## 0.0.2
 * Fixed podspec version